Rename and finish shorthand variable naming
