+++ bTagScript/block/loosevariablegetter.py
@@ -2,6 +2,7 @@ from typing import Optional
 
 from ..interface import Block
 from ..interpreter import Context
+from ..adapter import StringAdapter
 
 
 class LooseVariableGetterBlock(Block):
@@ -31,7 +32,11 @@ class LooseVariableGetterBlock(Block):
         return True
 
     def process(self, ctx: Context) -> Optional[str]:
-        if ctx.verb.declaration in ctx.response.variables:
+        if ctx.verb.declaration.startswith("$"):
+            varname = ctx.verb.declaration.split("$", 1)[1]
+            ctx.response.variables[varname] = StringAdapter(str(ctx.verb.payload))
+            return ""
+        elif ctx.verb.declaration in ctx.response.variables:
             return ctx.response.variables[ctx.verb.declaration].get_value(ctx.verb)
         else:
             return None
